requirements.txt
dnspython==2.1.0
pymongo==3.11.3
django-json==0.27
django==3.0.5
djongo==1.3.4
